+++ src/lib.rs
@@ -112,16 +112,17 @@ pub use packet::{
 pub use pes::{PStdBuffer, PesExtension, PesPacket, PesReassembler, ProgramPacketSequenceCounter};
 pub use psi::{
     decode_utc_time, iter_sections, mpeg2_crc32, BatTransportStream, BouquetAssociationTable,
-    ConditionalAccessTable, EitDateTime, EitDuration, EitEvent, EventInformationTable,
-    NetworkInformationTable, NitTransportStream, PmtStream, ProgramAssociationTable,
-    ProgramMapTable, PsiSectionAssembler, RstEntry, RunningStatus, RunningStatusTable, SdtService,
-    SectionIter, ServiceDescriptionTable, TimeDateTable, TimeOffsetTable,
-    TransportStreamDescriptionTable, BAT_PID, BAT_TABLE_ID, CAT_PID, CAT_TABLE_ID,
-    EIT_ACTUAL_PF_TABLE_ID, EIT_ACTUAL_SCHEDULE_FIRST, EIT_ACTUAL_SCHEDULE_LAST,
-    EIT_OTHER_PF_TABLE_ID, EIT_OTHER_SCHEDULE_FIRST, EIT_OTHER_SCHEDULE_LAST, EIT_PID,
-    MAX_PSI_SECTION_LEN, NIT_ACTUAL_TABLE_ID, NIT_OTHER_TABLE_ID, NIT_PID, PAT_PID, PAT_TABLE_ID,
-    PMT_TABLE_ID, RST_PID, RST_TABLE_ID, SDT_ACTUAL_TABLE_ID, SDT_OTHER_TABLE_ID, SDT_PID,
-    ST_TABLE_ID, TDT_TABLE_ID, TDT_TOT_PID, TOT_TABLE_ID, TSDT_PID, TSDT_TABLE_ID,
+    ConditionalAccessTable, DiscontinuityInformationTable, EitDateTime, EitDuration, EitEvent,
+    EventInformationTable, NetworkInformationTable, NitTransportStream, PmtStream,
+    ProgramAssociationTable, ProgramMapTable, PsiSectionAssembler, RstEntry, RunningStatus,
+    RunningStatusTable, SdtService, SectionIter, ServiceDescriptionTable, StuffingTable,
+    TimeDateTable, TimeOffsetTable, TransportStreamDescriptionTable, BAT_PID, BAT_TABLE_ID,
+    CAT_PID, CAT_TABLE_ID, DIT_PID, DIT_TABLE_ID, EIT_ACTUAL_PF_TABLE_ID,
+    EIT_ACTUAL_SCHEDULE_FIRST, EIT_ACTUAL_SCHEDULE_LAST, EIT_OTHER_PF_TABLE_ID,
+    EIT_OTHER_SCHEDULE_FIRST, EIT_OTHER_SCHEDULE_LAST, EIT_PID, MAX_PSI_SECTION_LEN,
+    NIT_ACTUAL_TABLE_ID, NIT_OTHER_TABLE_ID, NIT_PID, PAT_PID, PAT_TABLE_ID, PMT_TABLE_ID, RST_PID,
+    RST_TABLE_ID, SDT_ACTUAL_TABLE_ID, SDT_OTHER_TABLE_ID, SDT_PID, ST_TABLE_ID, TDT_TABLE_ID,
+    TDT_TOT_PID, TOT_TABLE_ID, TSDT_PID, TSDT_TABLE_ID,
 };
 pub use stream_type::StreamType;